requirements.txt
openai==1.3.5
hydra-core==1.3.2
omegaconf==2.3.0
pandas==2.1.4
seaborn==0.13.1
matplotlib==3.8.2